www/boc/boc/requirements.txt

MySQL-python==1.2.5
Pillow==2.6.1
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
bleach==1.4
certifi==14.05.14
django-allauth==0.23.0
django-app-namespace-template-loader==0.2
django-autoslug==1.8.0
django-bootstrap-form==3.1
django-ckeditor-updated==4.4.4
django-clever-selects==0.6.9
django-contrib-comments==1.6.1
django-crontab==0.6.0
django-easy-pjax==1.2.0
django-forms-bootstrap==3.0.1
django-jsonfield==0.9.13
django-localflavor==1.1
django-locality==0.2.2
django-material==0.4.1
django-mptt==0.6.1
django-stripe-payments==2.0.0
django-tinymce==1.5.3
django-xmlrpc==0.1.5
filebrowser-safe==0.3.6
future==0.9.0
gevent==1.0.1
gevent-socketio==0.2.2
gevent-websocket==0.9.3
grappelli-safe==0.3.13
greenlet==0.4.5
html5lib==0.999
lxml==3.4.2
oauthlib==1.0.3
pyparsing==2.0.3
python-openid==2.2.5
pytz==2014.10
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sphinx-me==0.3
stripe==1.21.0
tornado==4.1
torndb==0.3
tzlocal==1.0
wsgiref==0.1.2
